app/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.6.0
djangorestframework==3.16.1
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
unicorn==2.1.4
